requirements.txt
django==4.1.3
djangorestframework==3.14.0
django-filter==22.1
django-allauth==0.51.0






